fix(firmware): apply ultra-review findings to PanelGroup and ads1115_analog test
PanelGroup.cpp:
- Log warning when MAX_INT_PINS (8) exceeded instead of silently skipping
  attachInterrupt; affected chip falls back to 20 ms polling fallback
- Document _rxCount uint16_t rollover (~131 s at full bus load) — intentional
  for a diagnostic counter; comment prevents future widening without reason

Tests/PinRef/ads1115_analog.cpp:
- Cast upper bound of second-read tolerance check to uint32_t to prevent
  silent uint16_t overflow if val near 65535 (safe in practice at GAIN_ONE
  with 3.3 V, but formally correct)

+++ Firmware/Tests/PinRef/tests/ads1115_analog/ads1115_analog.cpp
@@ -100,8 +100,8 @@ void setup() {
     // Allow ±10% variation between two back-to-back reads
     uint16_t tolerance = val / 10 + 200;
     check("Second read within ±10% of first",
-          val2 >= (val > tolerance ? val - tolerance : 0) &&
-          val2 <= val + tolerance);
+          val2 >= (val > tolerance ? val - tolerance : 0u) &&
+          (uint32_t)val2 <= (uint32_t)val + tolerance);
 
     STM32Board::diagSerial().println(pass ? "=== ALL PASS ===" : "=== FAIL ===");
 }
